e2e/autoscaler: create MachineSets as part of test setup
Rather than use the MachineSet's that may (or may not) come with the
installed platform we now create the number of MachineSet's we
actually need as part of test setup; on AWS we get three and on Azure
we currently get one. To handle these kind of differences this commit
changes the autoscaler test to create what it needs as part of test
setup.

The new setup requires at least one MachineSet with an accompanying
Node to be present. From that MachineSet we create 3 new MachineSet's
solely for autoscaler testing. Labels and the ProviderSpec are
propagated to the new MachineSet's from the first MachineSet
discovered on the cluster.

The MachineSet is annotated with:

```yaml
apiVersion: machine.openshift.io/v1beta1
kind: MachineSet
metadata:
  annotations:
    machine.openshift.io/instance-cpu-capacity: "1"
    machine.openshift.io/instance-memory-capacity: 3840Mi
    machine.openshift.io/scale-from-zero: "true"
```

and the CPU and memory capacity are derived from an existing worker
node. Additionally we now scale the MachineSet's (via the associated
MachineAutoscalers) from 0. This helps to keep the test duration sane.

This change requires [1] which allows the cluster-autoscaler to scale
from 0.

[1] https://github.com/openshift/kubernetes-autoscaler/pull/110 which